Going on a tangent to add menu support.
Before getting into the robot stuff I need to create a support structure
for the GUI. I started on a basic menu component that has a theme and
constraints for display and location repectively. Some of the functions
will not be implemented at this stage because they are not yet needed.