Made headway with the input.
Not working but the direction I'm going in is to have the canvas ready to
take input, then when a scene is changed out the current scene is
registered with an input manager that adds whatever input is coming in to
a queue. The game object requests the manager to process input which does
so by passing it to the current scene if that scene is active. The scene
determines what to do based on the mapping agreement between it and the
manager.