Got audio working.
Audio has to be in WAV format I guess.
Modified input a bit to include the key code in the argument list when it calls
methods. It will also check through each key code in a set if it's listed in the
hold input hash. I may want to change how the arguments for audio is passed.
I think a hash would be easier.

Still don't care for the graphics object, need to look at alternatives but
its fine for images.

Next up is tiling and movement.