Now with multi-threading!
A thread is created with every update the scene calls on a group.
It then waits for the update threads to finish before returning
to the main process. To ensure safety between threads, objects
can only access objects in the same group. If an object must
affect a different group it should call a request to the scene.
After update finish the requests will be processed in order.

Everything has been tested in v1 and v2, littlegame.rb and test_littlegame.rb
(except for littleobjects.rb)
Assume everything else breaks

Next is to work on how the objects will interact between groups.